Renamed blockchain_statistics plugin to statsd plugin. Added changes to golosd and configs. #343
