+++ plugins/statsd/include/golos/plugins/statsd/plugin.hpp
@@ -17,17 +17,17 @@
 // various template automagic depends on them being known at compile
 // time.
 //
-#ifndef BLOCKCHAIN_STATISTICS_SPACE_ID
-#define BLOCKCHAIN_STATISTICS_SPACE_ID 9
+#ifndef STATSD_SPACE_ID
+#define STATSD_SPACE_ID 9
 #endif
 
-#ifndef BLOCKCHAIN_STATISTICS_PLUGIN_NAME
-#define BLOCKCHAIN_STATISTICS_PLUGIN_NAME "chain_stats"
+#ifndef STATSD_PLUGIN_NAME
+#define STATSD_PLUGIN_NAME "statsd"
 #endif
 
 namespace golos {
 namespace plugins {
-namespace blockchain_statistics {
+namespace statsd {
 
 using namespace golos::chain;
 using boost::program_options::options_description;
@@ -37,7 +37,7 @@ using boost::program_options::variables_map;
 class plugin final : public appbase::plugin<plugin> {
 public:
     static const std::string &name() {
-        static std::string name = BLOCKCHAIN_STATISTICS_PLUGIN_NAME;
+        static std::string name = STATSD_PLUGIN_NAME;
         return name;
     }
 
@@ -65,4 +65,4 @@ private:
     std::unique_ptr<plugin_impl> _my;
 };
 
-} } } // golos::plugins::blockchain_statistics
+} } } // golos::plugins::statsd
