+++ plugins/statsd/include/golos/plugins/statsd/statistics_sender.hpp
@@ -4,7 +4,7 @@
 #include <boost/asio/ip/udp.hpp>
 #include <fc/uint128_t.hpp>
 #include <golos/chain/steem_object_types.hpp>
-#include <golos/plugins/blockchain_statistics/runtime_bucket_object.hpp>
+#include <golos/plugins/statsd/runtime_bucket_object.hpp>
 
 using namespace golos::chain;
 
@@ -26,8 +26,8 @@ public:
     /// returns statistics recievers endpoints
     std::vector<std::string> get_endpoint_string_vector();
 
-    golos::plugins::blockchain_statistics::runtime_bucket_object previous_bucket;
-    golos::plugins::blockchain_statistics::runtime_bucket_object current_bucket;
+    golos::plugins::statsd::runtime_bucket_object previous_bucket;
+    golos::plugins::statsd::runtime_bucket_object current_bucket;
     bool is_previous_bucket_set;
 private:
     // Stat sender will send data to all endpoints from recipient_endpoint_set
