+++ programs/golosd/main.cpp
@@ -13,7 +13,7 @@
 #include <golos/plugins/test_api/test_api_plugin.hpp>
 #include <golos/plugins/social_network/social_network.hpp>
 #include <golos/plugins/account_history/plugin.hpp>
-#include <golos/plugins/blockchain_statistics/plugin.hpp>
+#include <golos/plugins/statsd/plugin.hpp>
 #include <golos/plugins/account_by_key/account_by_key_plugin.hpp>
 #include <golos/plugins/private_message/private_message_plugin.hpp>
 #include <golos/plugins/auth_util/plugin.hpp>
@@ -63,7 +63,7 @@ namespace golos {
             appbase::app().register_plugin<golos::plugins::test_api::test_api_plugin>();
             appbase::app().register_plugin<golos::plugins::market_history::market_history_plugin>();
             appbase::app().register_plugin<golos::plugins::account_history::plugin>();
-            appbase::app().register_plugin<golos::plugins::blockchain_statistics::plugin>();
+            appbase::app().register_plugin<golos::plugins::statsd::plugin>();
             appbase::app().register_plugin<golos::plugins::account_by_key::account_by_key_plugin>();
             appbase::app().register_plugin<golos::plugins::private_message::private_message_plugin>();
             appbase::app().register_plugin<golos::plugins::auth_util::plugin>();
