+++ share/golosd/config/config.ini
@@ -35,7 +35,7 @@ rpc-endpoint = 0.0.0.0:2001
 
 shared-file-size = 64G
 
-plugin = chain p2p json_rpc webserver network_broadcast_api witness test_api database_api private_message follow social_network market_history account_by_key account_history chain_stats block_info raw_block debug_node
+plugin = chain p2p json_rpc webserver network_broadcast_api witness test_api database_api private_message follow social_network market_history account_by_key account_history statsd block_info raw_block debug_node
 
 
 # JSON list of [nblocks,nseconds] pairs, see documentation/bcd-trigger.md
